Move sleep util to shared folder
